+codestyle 9

+++ tasks/mpi/kovalev_k_bubble_sort_oddeven_transposition/src/sourse.cpp
@@ -24,7 +24,7 @@ bool kovalev_k_bubble_sort_oddeven_transposition_mpi::BubbleSortOddEvenTransposi
       res.clear();
       for (size_t i = 0; i < loc_v.size(); i++) tmp.push_back(loc_v[i]);
       size_t iter1 = 0, iter2 = sendcounts[partner];
-      while (iter2 < tmp.size() || iter1 < sendcounts[partner]) {
+      while (iter2 < tmp.size() || iter1 < static_cast<size_t>(sendcounts[partner])) {
         if ((iter1 < static_cast<size_t>(sendcounts[partner]) && iter2 < tmp.size() && tmp[iter1] <= tmp[iter2]) ||
             (iter1 < static_cast<size_t>(sendcounts[partner]) && iter2 == tmp.size())) {
           res.push_back(tmp[iter1]);
